Making an hook for menu dropdown
